+++ RenderingEngine/src/rendering/shaders/ShaderProgram.cpp
@@ -6,7 +6,7 @@ ShaderProgram::ShaderProgram(ID3D11Device* device, const VertexShader& vertexSha
 	: vertexShader{vertexShader_.shader},
 	  pixelShader{pixelShader_.shader}
 {
-	const HRESULT result = device->CreateInputLayout(layout_.data(), layout_.size(),
+	const HRESULT result = device->CreateInputLayout(layout_.data(), static_cast<UINT>(layout_.size()),
 	                                                 vertexShader_.bytecode->GetBufferPointer(),
 	                                                 vertexShader_.bytecode->GetBufferSize(),
 	                                                 &inputLayout);
